split WorkspaceSpec
